Rename `bpm-test-agent` to `test-server`
 * avoid colliding with `bpm` job name
 * makes purpose of job clear
