--- jobs/bpm-test-agent/monit
@@ -1,11 +1,0 @@
-check process bpm-test-agent
-  with pidfile /var/vcap/sys/run/bpm/bpm-test-agent/bpm-test-agent.pid
-  start program "/var/vcap/jobs/bpm/bin/bpm start bpm-test-agent"
-  stop program "/var/vcap/jobs/bpm/bin/bpm stop bpm-test-agent"
-  group vcap
-
-check process alt-bpm-test-agent
-  with pidfile /var/vcap/sys/run/bpm/bpm-test-agent/alt-server.pid
-  start program "/var/vcap/jobs/bpm/bin/bpm start bpm-test-agent -p alt-server"
-  stop program "/var/vcap/jobs/bpm/bin/bpm stop bpm-test-agent -p alt-server"
-  group vcap
