--- packages/bpm-test-agent/spec
@@ -1,11 +1,0 @@
----
-name: bpm-test-agent
-
-dependencies:
-  - golang
-
-files:
-  - bpm-acceptance/fixtures/bpm-test-agent/*.go # gosub
-  - bpm-acceptance/fixtures/bpm-test-agent/handlers/*.go # gosub
-  - golang.org/x/sys/unix/*.go # gosub
-  - golang.org/x/sys/unix/*.s # gosub
