+++ packages/test-server/packaging
@@ -7,7 +7,7 @@ export GOPATH=$BOSH_INSTALL_TARGET
 export GOROOT=$(readlink -nf /var/vcap/packages/golang)
 export PATH=$GOROOT/bin:$PATH
 
-go install bpm-acceptance/fixtures/bpm-test-agent
+go install bpm-acceptance/fixtures/test-server
 
 # clean up source artifacts
 rm -rf ${BOSH_INSTALL_TARGET}/src ${BOSH_INSTALL_TARGET}/pkg
