+++ packages/test-server/spec
@@ -1,0 +1,11 @@
+---
+name: test-server
+
+dependencies:
+  - golang
+
+files:
+  - bpm-acceptance/fixtures/test-server/*.go # gosub
+  - bpm-acceptance/fixtures/test-server/handlers/*.go # gosub
+  - golang.org/x/sys/unix/*.go # gosub
+  - golang.org/x/sys/unix/*.s # gosub
