+++ src/bpm-acceptance/fixtures/test-server/handlers/curl.go
