+++ src/bpm-acceptance/fixtures/test-server/handlers/masked_paths.go
